Move results retrieval to TeamResultsAction
That means it is no longer necessary to trigger the build via
/team-build/ to be able to download collected results.
It also means the ZIP can be downloaded from
/job/JOB_NAME/BUILD_NUMBER/team-results/zip
